fixed command issues with interactionCreate.js
it now checks if you are using a specific command or any other command.

created a new command called support, calls to a api with the users message and returns the results in an embedded format for discord.

need to program in a configuration file, so it is easier to change parameters without editing source code.

will also need to use search parameter based on file location, so if you wanted to say, create a new command/child of current command, it would be based on current folder name as config string search parameter.